# Conflicts:
#	Casper.Network.SDK/JsonRpc/ResultTypes/GetBlockResult.cs
#	Casper.Network.SDK/JsonRpc/ResultTypes/GetBlockTransfersResult.cs
#	Casper.Network.SDK/JsonRpc/ResultTypes/QueryGlobalStateResult.cs
#	Casper.Network.SDK/SSE/BlockAdded.cs
#	Casper.Network.SDK/Types/Block.cs
#	Casper.Network.SDK/Types/EraEnd.cs
#	Casper.Network.SDK/Types/GlobalStateKey/GlobalStateKey.cs
#	Casper.Network.SDK/Types/Transfer.cs

+++ Casper.Network.SDK/Types/StateIdentifier.cs
@@ -6,7 +6,7 @@ namespace Casper.Network.SDK.Types
     {
         private string _stateRootHash;
         private string _blockHash;
-        private int? _blockHeight;
+        private ulong? _blockHeight;
 
         private StateIdentifier()
         {
@@ -32,7 +32,7 @@ namespace Casper.Network.SDK.Types
             };
         }
         
-        public static StateIdentifier WithBlockHeight(int blockHeight)
+        public static StateIdentifier WithBlockHeight(ulong blockHeight)
         {
             return new StateIdentifier
             {
